update pass to transform OtherNode to LoadVariableNode
Any unknown token is first parsed to a OtherNode.
In a later pass any OtherNode which has a name that is somewhere
defined in a DefineWordNode is transformed to a CallWordNode.
An OtherNode which has a name that is defined as a Variable is
transformed to a LoadVariable node, where the variable is used again.