parse default struct values and store them in StructMember
A struct can have default values like this:
struct Player {
  age = 10
  movement_speed = 1
}
These default values are parsed and stored in a StructMember, which
defines a struct member with its name and default string value. To
output code to use, these string values have to be used in a further
pass to emit code at the beginning of the program to use the default
value as either a number or address and set it with setters to the
corresponding struct.